feat: update bootstrap script and add shared SSH configuration for CAGELAB systems

+++ bootstrap.sh
@@ -5,6 +5,12 @@ export SPATH="$HOME/Code/Setup"
 uname -a | grep -iq "microsoft" && MOD="WSL"
 uname -a | grep -iq "aarch64|armv7" && MOD="RPi"
 PLATFORM=$(uname -s)$MOD
+mversion="R2025b"
+if [[ $PLATFORM == "Darwin" ]]; then
+	mpath="/Applications/MATLAB_$mversion.app/bin/matlab"
+else
+	mpath="/usr/local/MATLAB/$mversion/bin/matlab"
+fi
 
 sudo apt install git
 sudo apt install curl
@@ -36,10 +42,10 @@ sudo ln -svf "$SPATH/config/toggleInput" /usr/local/bin/toggleInput &&
 	systemctl --user enable toggleInput.service &&
 	systemctl --user start toggleInput.service
 
-# Install X-CMD
+# Install X-CMD (cross-platform script support)
 [[ ! -d $HOME/.x-cmd.root ]] && eval "$(curl https://get.x-cmd.com)"
 
-# Install Pixi
+# Install Pixi (cross-platform package manager)
 [[ ! -d $HOME/.pixi/bin ]] && eval curl -fsSL https://pixi.sh/install.sh | bash
 
 # Install Netbird
@@ -55,20 +61,20 @@ fi
 printf "Shall we use MPM to get MATLAB? [y / n]:  "
 read -r ans
 if [[ $ans == 'y' ]]; then
-	version="R2025a"
 	products='MATLAB Curve_Fitting_Toolbox Instrument_Control_Toolbox MATLAB_Report_Generator Optimization_Toolbox Parallel_Computing_Toolbox Signal_Processing_Toolbox Statistics_and_Machine_Learning_Toolbox'
 	curl -L -o /usr/local/bin/mpm https://www.mathworks.com/mpm/glnxa64/mpm
 	chmod +x /usr/local/bin/mpm
 	sudo mkdir -p /usr/local/MATLAB &&
 	sudo chown "$USER":"$USER" /usr/local/MATLAB &&
 	sudo chmod 777 /usr/local/MATLAB &&
-	/usr/local/bin/mpm install --release=$version --products=$products
+	/usr/local/bin/mpm install --release=$mversion --products=$products
 fi
+[[ -f $mpath ]] && ln -sfv /usr/local/MATLAB/$mversion/bin/matlab /usr/local/bin
 
 # APT + snap + flatpak packages
 if [ "$PLATFORM" = "Linux" ]; then
 	sudo apt --fix-broken install
-	sudo apt update && sudo apt -y upgrade
+	sudo apt update && sudo apt -y full-upgrade
 	sudo apt -my install apt-transport-https ca-certificates software-properties-common
 	sudo apt -my install build-essential zsh git gparted vim curl file mc
 	sudo apt -my install gawk mesa-utils exfatprogs
@@ -84,8 +90,10 @@ if [ "$PLATFORM" = "Linux" ]; then
 	sudo apt -my install openjdk-17-jre
 	sudo apt -my install flatpak
 
-	sudo snap install --classic code
+	# Install some snap packages
+	[[ ! -f $(which code) ]] && sudo snap install --classic code
 
+	# flatpak
 	flatpak remote-add --if-not-exists flathub https://dl.flathub.org/repo/flathub.flatpakrepo
 	flatpak install -y flathub com.obsproject.Studio
 	
@@ -107,7 +115,7 @@ curl -sS https://raw.githubusercontent.com/Ccccraz/cogmoteGO/main/install.sh | s
 [[ ! -f /usr/bin/sunshine ]] && eget LizardByte/Sunshine -a '24.04' --to=./sunshine.deb && sudo dpkg -i ./sunshine.deb
 [[ ! -f /usr/bin/rotz ]] && eget volllly/rotz --to=/usr/local/bin
 
-# Clone our core repos
+# Clone our core repos from gitee
 mkdir -p "$HOME/Code"
 cd "$HOME/Code" || exit
 [[ ! -d 'Setup' ]] && git clone --recurse-submodules https://gitee.com/CogPlatform/Setup.git
@@ -126,18 +134,24 @@ cd ~ || exit
 	sudo ln -svf /usr/lib/x86_64-linux-gnu/libglut.so.3.12.0 /usr/lib/x86_64-linux-gnu/libglut.so.3
 
 # Setup PTB and opticka path:
-#cd "$HOME/Code/Psychtoolbox" || exit
-#$mpath/matlab -nodesktop -nosplash -r "SetupPsychToolbox; pause(1); cd ../../opticka; addOptickaToPath; pause(1); exit"
-
-cd "$HOME" || exit
+if [[ -f $mpath ]]; then
+	printf "Shall we setup MATLAB path? [y / n]:  "
+	read -r ans
+	if [[ $ans == 'y' ]]; then
+		cd "$HOME/Code/Psychtoolbox" || exit
+		matlab -nodesktop -nosplash -r "SetupPsychToolbox; pause(1); cd ../../opticka; addOptickaToPath; pause(1); exit"
+	fi
+	cd "$HOME" || exit
+fi
 
 # Link pixi-global.toml
 rm -rf $HOME/.pixi/manifests
 mkdir -p $HOME/.pixi/manifests && ln -svf $SPATH/config/pixi-global.toml $HOME/.pixi/manifests/
 [[ -f $(which pixi) ]] && pixi global sync
 
-# link some cagelab scripts
+# link some cagelab stuff
 ln -svf "$HOME/Code/CageLab/software/scripts/"* "$HOME/bin"
+ln -svf "$HOME/Code/Setup/config/sshconfig" "$HOME/.ssh/config"
 
 # Link .zshrc
 [[ -e ~/.zshrc ]] && cp ~/.zshrc ~/.zshrc"$(date -Iseconds)".bak
